+++ src/main/java/org/teacon/powertool/inspection/property/NumberProperty.java
@@ -2,9 +2,9 @@ package org.teacon.powertool.inspection.property;
 
 public interface NumberProperty<T extends Number> extends Property<T> {
 
-    void setNumber(float value);
+    void setNumber(double value);
 
-    float getNumber();
+    double getNumber();
 
     @Override
     default void set(T value) {
